refactor(pi): extract launcher support module
Agent-Model: gpt-5
Agent-Family: openai
Agent-Session: 019e6c4d-7afb-7ac0-ac79-4582d0e36455
Agent-Step: 0.1.3
